Remove SettlorIndividualExtractor (this is now combined within LivingSettlorExtractor), update unit tests
